break out type conversion module
